Encapsulate serde based doctest.
One of the tests included with the README is heavily based
on the support of serde. Each run of `cargo test` without
that feature would fail the test.

This change adds a module to encapsulte the test and
ignore it as long the feature is missing. This also
required to add a custom main function to the test.